backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
requests
beautifulsoup4
apscheduler
scrapy
python-dotenv
